+++ src/friso_API.h
@@ -14,7 +14,14 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#ifdef _WIN32
+//yat, just take it as this way, 99 percent you will find no problem
+#if ( defined(_WIN32) || defined(_WINDOWS_) || defined(__WINDOWS_) )
+#	define FRISO_WINNT
+#else
+#	define FRISO_LINUX
+#endif
+
+#ifdef FRISO_WINNT
 #	define FRISO_API extern __declspec(dllexport)
 #	define __STATIC_API__ static
 #else